+++ src/briscola.rs
@@ -1,7 +1,7 @@
 use super::primitives::*;
 use std::collections::hash_map::*;
-use rand::seq::SliceRandom;
 use itertools::Itertools;
+use super::utils;
 
 #[derive(Debug)]
 pub struct Briscola {
@@ -18,21 +18,6 @@ pub struct Briscola {
     deck: CardDeck,
     briscola: CardSuit,
     next_player: Option<Player>,
-}
-
-impl Briscola {
-    fn get_card_name(card: &Card) -> String {
-        let c_type = match card.0 {
-            CardType::Jack => "Fante".to_owned(),
-            CardType::Queen => "Cavallo".to_owned(),
-            CardType::King => "Re".to_owned(),
-            CardType::Numeric(x) => match x {
-                1 => "Asso".to_owned(),
-                _ => format!("{}", x)
-            }
-        };
-        format!("{} di {}", c_type, String::from(&card.1))
-    }
 }
 
 impl Game for Briscola {
@@ -70,26 +55,7 @@ impl Game for Briscola {
         }
     }
     fn init() -> Self {
-        let mut rng = rand::thread_rng();
-        let mut deck_raw: Vec<u8> = (0..40).collect();
-        deck_raw.shuffle(&mut rng);
-        let deck: Vec<Card> = deck_raw.iter().map(|x| {
-            let value: u8 = x % 10;
-            // Ricordarsi di aggiungere 1
-            let suit = match x - value {
-                0 => CardSuit::Bastoni,
-                10 => CardSuit::Coppe,
-                20 => CardSuit::Denari,
-                _ => CardSuit::Spade
-            };
-            let c_type = match value {
-                9 => CardType::King,
-                8 => CardType::Queen,
-                7 => CardType::Jack,
-                _ => CardType::Numeric(value + 1)
-            };
-            (c_type, suit)
-        }).collect();
+        let deck = utils::random_deck(CardDeckType::Briscola);
         let mut teams = Vec::new();
         teams.push(Vec::new());
         teams.push(Vec::new());
@@ -241,7 +207,7 @@ impl Game for Briscola {
             self.get_scores().iter().enumerate().map(|x| format!("Team {}: {} punti", x.0, (x.1).1)).join("\n"),
             String::from(&self.briscola),
             self.get_next_player().map(|x| x.name).unwrap_or("".to_owned()),
-            self.table.iter().map(|x| format!("- {} ({})", Self::get_card_name(&x.1), x.0.name)).join("\n")
+            self.table.iter().map(|x| format!("- {} ({})", utils::get_card_name(&x.1), x.0.name)).join("\n")
         )
     }
 }
