+++ src/main.rs
@@ -1,5 +1,7 @@
 mod primitives;
 mod briscola;
+mod beccaccino;
+mod utils;
 use primitives::Game;
 use text_io::read;
